Move SearchIndexItem, SearchIndexItems to toolkit.utils
