+++ src/jdk.javadoc/share/classes/jdk/javadoc/internal/doclets/toolkit/util/IndexItem.java
@@ -25,8 +25,6 @@
 
 package jdk.javadoc.internal.doclets.toolkit.util;
 
-import jdk.javadoc.internal.doclets.formats.html.SearchIndexItem;
-
 import javax.lang.model.element.Element;
 import javax.lang.model.element.ExecutableElement;
 import javax.lang.model.element.ModuleElement;
